Move history to plugins folder, de-lint and es6ify
